Initial commit for rewriting entire server-side code
This repo is in the process of being completely revamped on both
client side and server side. This project was created when I didn't
know much about Javascript and it isn't code I'm happy with.

Main changes:
- Requests to Spotify API are done in the back end to keep the
authorization and refresh tokens hidden from the client
- App will no longer require user to manually authorize Spotify.
Instead, authorization is already done using my Spotify account using
the authorization code flow and I supply the refresh token to the
server side. If the token has expired, the access token is
automatically refreshed, so the user doesn't have to do anything